Add test for verify functions
