+++ be1-go/channel/lao/mod_test.go
@@ -1,14 +1,10 @@
 package lao
 
 import (
-	"encoding/json"
 	"fmt"
 	"io"
-	"os"
-	"path/filepath"
 	"popstellar/channel"
 	"popstellar/crypto"
-	"popstellar/message/messagedata"
 	"popstellar/message/query/method"
 	"popstellar/message/query/method/message"
 	"popstellar/network/socket"
@@ -96,108 +92,6 @@ func TestBaseChannel_ConsensusIsCreated(t *testing.T) {
 	consensusID := "channel0/consensus"
 	consensus := fakeHub.channelByID[consensusID]
 	require.NotNil(t, consensus)
-}
-
-func Test_Verify_Functions(t *testing.T) {
-	// Create the hub
-	keypair := generateKeyPair(t)
-
-	fakeHub, err := NewfakeHub(keypair.public, nolog, nil)
-	require.NoError(t, err)
-
-	// Create the channel
-	numMessages := 1
-
-	messages := make([]message.Message, numMessages)
-
-	channel := NewChannel("fzJSZjKf-2cbXH7kds9H8NORuuFIRLkevJlN7qQemjo=", fakeHub, messages[0], nolog)
-
-	laoChannel, ok := channel.(*Channel)
-	require.True(t, ok)
-
-	// Get the JSON
-	relativeExamplePath := filepath.Join("..", "..", "..", "protocol",
-		"examples", "messageData")
-
-	// Test LaoState
-	file := filepath.Join(relativeExamplePath, "lao_state.json")
-
-	buf, err := os.ReadFile(file)
-	require.NoError(t, err)
-
-	object, action, err := messagedata.GetObjectAndAction(buf)
-	require.NoError(t, err)
-
-	require.Equal(t, "lao", object)
-	require.Equal(t, "state", action)
-
-	var laoState messagedata.LaoState
-
-	err = json.Unmarshal(buf, &laoState)
-	require.NoError(t, err)
-
-	err = laoChannel.verifyMessageLaoState(laoState)
-	require.NoError(t, err)
-
-	// Test RollCallCreate
-	file = filepath.Join(relativeExamplePath, "roll_call_create.json")
-
-	buf, err = os.ReadFile(file)
-	require.NoError(t, err)
-
-	object, action, err = messagedata.GetObjectAndAction(buf)
-	require.NoError(t, err)
-
-	require.Equal(t, "roll_call", object)
-	require.Equal(t, "create", action)
-
-	var rollCallCreate messagedata.RollCallCreate
-
-	err = json.Unmarshal(buf, &rollCallCreate)
-	require.NoError(t, err)
-
-	err = laoChannel.verifyMessageRollCallCreate(rollCallCreate)
-	require.NoError(t, err)
-
-	// Test RollCallOpen
-	file = filepath.Join(relativeExamplePath, "roll_call_open.json")
-
-	buf, err = os.ReadFile(file)
-	require.NoError(t, err)
-
-	object, action, err = messagedata.GetObjectAndAction(buf)
-	require.NoError(t, err)
-
-	require.Equal(t, "roll_call", object)
-	require.Equal(t, "open", action)
-
-	var rollCallOpen messagedata.RollCallOpen
-
-	err = json.Unmarshal(buf, &rollCallOpen)
-	require.NoError(t, err)
-
-	err = laoChannel.verifyMessageRollCallOpen(rollCallOpen)
-	require.NoError(t, err)
-
-	// Test RollCallClose
-	file = filepath.Join(relativeExamplePath, "roll_call_close.json")
-
-	buf, err = os.ReadFile(file)
-	require.NoError(t, err)
-
-	object, action, err = messagedata.GetObjectAndAction(buf)
-	require.NoError(t, err)
-
-	require.Equal(t, "roll_call", object)
-	require.Equal(t, "close", action)
-
-	var rollCallClose messagedata.RollCallClose
-
-	err = json.Unmarshal(buf, &rollCallClose)
-	require.NoError(t, err)
-
-	err = laoChannel.verifyMessageRollCallClose(rollCallClose)
-	require.NoError(t, err)
 }
 
 // -----------------------------------------------------------------------------
